修改附件管理的删除错误 增加了组件页面帮助； 文件附件和avatar的CURD错误

+++ fast-framework/src/main/java/com/j2eefast/framework/utils/FileUploadUtils.java
@@ -102,10 +102,11 @@ public class FileUploadUtils {
 	public void deleteFileRelation(Long fileId) {
 		SysFilesEntity fileEntity = ConstantFactory.me().getSysFileService().getById(fileId);
 		if (null != fileEntity) {
-			String filePath = (fileEntity.getClassify().equals("0") || fileEntity.getClassify().equals("1")) ? Global.getAttachPath() : Global.getEditorPath() + fileEntity.getFilePath();
+			String divPath = (fileEntity.getClassify().equals("0") || fileEntity.getClassify().equals("1")) ? Global.getAttachPath() : Global.getEditorPath() ;
+			String filePath = divPath + fileEntity.getFilePath() ;
 			File file = new File(filePath);
 			//删除文件
-			if (file.exists()) {
+			if (file.exists() && file.isFile()) {
 				FileUtils.delete(new File(filePath));
 			}
 			
@@ -139,76 +140,99 @@ public class FileUploadUtils {
      * @param imgName 页面控件Name
      * @return
      */
+	@Deprecated
     public static String getAvatarImg(String imgName){
         HttpServletRequest request = HttpContextUtil.getRequest();
         String imgBase64 = request.getParameter(imgName);
         return saveImgBase64(imgBase64);
     }
 
-    /**
+     /**
      * 图片裁剪保存图片信息
      * @param imgBase64
      * @return 返回裁剪控件图片存放相对路径
      */
-    
+	@Deprecated
     public static String saveImgBase64(String imgBase64) {
-    	return saveImgBase64( imgBase64,null)   ; 	
+    	return saveImgBase64(imgBase64,null,null,null)   ; 	
     }
     
-    public static String saveImgBase64(String imgBase64,String bizType) {
-    	 if (imgBase64.equals(StrUtil.EMPTY)) {
-             return null;
-         }
+    
+    /**
+    * @Title: saveImgBase64 
+    * @Description: 图片裁剪保存图片信息，当一个业务ID,包括多个avatar属性时，需要传入文件名，本方法在业务实体Save 方法后使用
+    * @param imgBase64  
+    * @param bizType  业务类型
+    * @param bizId   业务ID
+    * @param fileName  为空时,使用bizId+后缀名 命名
+    * @return  String  保文件的路径名
+    * @author mfksn001@163.com
+    * @Date: 2020年8月24日
+     */
+    public static String saveImgBase64(String imgBase64,String bizType,Long bizId,String fileName) {
+    	if (imgBase64.equals(StrUtil.EMPTY)) {
+            return null;
+        }
 
-         String extension = null;
-         String type = StrUtil.subBetween(imgBase64, "data:", ";base64,");
-         if (StrUtil.containsIgnoreCase(type, "image/jpeg")) {
-             extension = "jpg";
-         } else if (StrUtil.containsIgnoreCase(type, "image/gif")) {
-             extension = "gif";
-         } else {
-             extension = "png";
-         }
-         
-         
-         String fileName = IdUtil.fastSimpleUUID() + "." + extension; 
+        String extension = null;
+        String type = StrUtil.subBetween(imgBase64, "data:", ";base64,");
+        if (StrUtil.containsIgnoreCase(type, "image/jpeg")) {
+            extension = "jpg";
+        } else if (StrUtil.containsIgnoreCase(type, "image/gif")) {
+            extension = "gif";
+        } else {
+            extension = "png";
+        }
+        String saveName = String.valueOf(bizId)+ "." + extension; ;
+        if (StringUtils.isNotBlank(fileName)) {
+        	//String saveName = IdUtil.fastSimpleUUID() + "." + extension; 
+        	saveName = fileName + "." + extension; 
+        }
+        
 
-         String imageUrl = File.separator + "avatar"
-                 + File.separator +
-                 DateFormatUtils.format(new Date(), "yyyy/MM/dd") + File.separator + fileName ;
+        String imageUrl = File.separator + "avatar"
+                + File.separator +
+                DateFormatUtils.format(new Date(), "yyyy/MM/dd") + File.separator + saveName ;
 
+        String base64 = StrUtil.subAfter(imgBase64, "base64,", true);
+        if (StrUtil.isBlank(base64)) {
+            return null;
+        }
 
-         String base64 = StrUtil.subAfter(imgBase64, "base64,", true);
-         if (StrUtil.isBlank(base64)) {
-             return null;
-         }
+        byte[] data = Base64.decode(base64);
 
-         byte[] data = Base64.decode(base64);
-
-         try {
-         	String fullPath =  Global.getAttachPath() + imageUrl ;
-             FileUtil.writeBytes(data, Global.getAttachPath() + imageUrl);
-             SysFilesEntity sysFile = new SysFilesEntity();
-             sysFile.setFileMd5(ToolUtil.encodingFilename(fileName));
-             sysFile.setFileName(fileName);
-             sysFile.setFilePath(imageUrl);
-             sysFile.setClassify("1");
-             sysFile.setFileSize(new BigDecimal(FileUtil.size(new File(fullPath))));
-             ConstantFactory.me().getSysFileService().save(sysFile);
-             if (StringUtils.isNotBlank(bizType)) {
+        try {
+        	String fullPath =  Global.getAttachPath() + imageUrl ;
+            FileUtil.writeBytes(data, Global.getAttachPath() + imageUrl);
+            SysFilesEntity sysFile = new SysFilesEntity();
+            sysFile.setFileMd5(ToolUtil.encodingFilename(saveName));
+            sysFile.setFileName(saveName);
+            sysFile.setFilePath(imageUrl);
+            sysFile.setClassify("1");
+            sysFile.setFileSize(new BigDecimal(FileUtil.size(new File(fullPath))));
+            ConstantFactory.me().getSysFileService().save(sysFile);
+            if (StringUtils.isNotBlank(bizType)) {
 	             SysFileUploadEntity sysFileUploadEntity  = new SysFileUploadEntity();
-	             //sysFileUploadEntity.setBizId(fileName);
-	             sysFileUploadEntity.setFileName(fileName);
+	             sysFileUploadEntity.setBizId(bizId);
+	             sysFileUploadEntity.setFileName(saveName);
 	             sysFileUploadEntity.setFileId(sysFile.getId());
 	             sysFileUploadEntity.setBizType(bizType);
-                 ConstantFactory.me().getSysFileUploadService().saveSysFileUpload(sysFileUploadEntity);
-             }
-         } catch (IORuntimeException e) {
-             e.printStackTrace();
-         }
+                ConstantFactory.me().getSysFileUploadService().saveSysFileUpload(sysFileUploadEntity);
+            }
+        } catch (IORuntimeException e) {
+            e.printStackTrace();
+        }
 
-         return imageUrl;
+        return imageUrl;
     }
-
+    /*
+     * 使用业务ID作为文件名
+     */
+    public static String saveImgBase64(String imgBase64,String bizType,Long bizId) {
+       return saveImgBase64( imgBase64, bizType, bizId,null);
+    }
+    
+ 
+ 
     
 }
